Day 9 hacky

+++ src/util/Pipe.scala
@@ -3,13 +3,13 @@ package util
 import java.util.concurrent.{BlockingQueue, LinkedBlockingQueue}
 
 class Pipe {
-  val data: BlockingQueue[Int] = new LinkedBlockingQueue[Int]()
+  val data: BlockingQueue[BigInt] = new LinkedBlockingQueue[BigInt]()
 
   def empty(): Boolean = data.isEmpty
   def size(): Int = data.size
-  def write(v: Int): Unit = data.add(v)
-  def read(): Int = data.take()
-  def readLast(): Int = {
+  def write(v: BigInt): Unit = data.add(v)
+  def read(): BigInt = data.take()
+  def readLast(): BigInt = {
     while(data.size() > 1 ) {
       data.take()
     }
